+++ src/utils/generalSettings.ts
@@ -4,7 +4,7 @@ export const countriesCarouselSettings = {
     speed: 500,
     slidesToShow: 11,
     slidesToScroll: 1,
-    arrows: true,
+    arrows: false,
     swipeToSlide: true,
     responsive: [
       {
